+++ code/modules/overmap/ships/computers/engine_control.dm
@@ -30,6 +30,7 @@
 	var/data = list()
 	var/mob/living/silicon/silicon = user
 	data["viewing_silicon"] = ismachinerestricted(silicon)
+	data["linked"] = linked
 	data["global_state"] = linked.engines_state
 	data["global_limit"] = round(linked.thrust_limit*100)
 	var/total_thrust = 0
@@ -52,9 +53,7 @@
 	return data
 
 /obj/machinery/computer/ship/engines/tgui_act(action, list/params)
-	. = ..()
-	if(.)
-		return
+	UI_ACT_CHECK
 
 	.=TRUE
 	add_fingerprint(usr)
@@ -93,6 +92,9 @@
 			var/datum/ship_engine/E = locate(params["ref"])
 			if(istype(E))
 				E.toggle()
+			return TOPIC_REFRESH
+		if("sync")
+			sync_linked(usr)
 			return TOPIC_REFRESH
 
 	return TOPIC_NOACTION
