+++ code/modules/overmap/ships/computers/ship.dm
@@ -21,7 +21,9 @@ somewhere on that shuttle. Subtypes of these can be then used to perform ship ov
 	. = attempt_hook_up_recursive(sector)
 	if(. && linked && user)
 		to_chat(user, "<span class='notice'>[src] reconnected to [linked]</span>")
-		user << browse(null, "window=[src]") // close reconnect dialog
+	else
+		to_chat(usr,"<span class='warning'>No shuttles near.</span>")
+
 /obj/machinery/computer/ship/proc/attempt_hook_up_recursive(obj/effect/overmap/visitable/ship/sector)
 	if(attempt_hook_up(sector))
 		return sector
