+++ code/modules/shuttles/shuttle_console.dm
@@ -109,8 +109,7 @@
 
 
 /obj/machinery/computer/shuttle_control/proc/handle_topic_href(var/datum/shuttle/autodock/shuttle, action, list/params)
-	if(.)
-		return
+	UI_ACT_CHECK
 
 	.=TRUE
